pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "b12-submit"
version = "0.1.0"
description = "Submit job application payload to B12"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
b12-submit = "b12_submit.__main__:main"
